+++ src/rtcmix/sfcopy.c
@@ -15,7 +15,7 @@ extern SFHEADER      sfdesc[NFILES];
 
 
 double
-sfcopy(float p[], int n_args)
+sfcopy(double p[], int n_args)
 {
 	int maxread,n,input,output,bytes,jj;
